feat: rework whole app to comply with expo new structure
