+++ app/game.tsx
@@ -1,0 +1,5 @@
+import Game from '@/components/Game';
+
+export default function StackLayout() {
+  return <Game />;
+}
