+++ app/settings.tsx
@@ -1,0 +1,5 @@
+import Settings from '@/components/Settings';
+
+export default function StackLayout() {
+  return <Settings />;
+}
