+++ hooks/useGameManager.ts
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+
+import { GameManagerContext } from '@/providers/GameManagerProvider';
+
+export default function useGameManager() {
+  return useContext(GameManagerContext);
+}
