+++ hooks/useTimerManager.ts
@@ -1,7 +1,6 @@
 import { useContext } from 'react';
 
-import { TimerManagerContext } from '../providers';
-
+import { TimerManagerContext } from '@/providers/TimerManagerProvider';
 
 export default function useTimerManager() {
     return useContext(TimerManagerContext);
