+++ package.json
@@ -1,40 +1,50 @@
 {
   "name": "minesweeper",
-  "version": "2.1.1",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "expo-router/entry",
+  "version": "1.0.0",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
-    "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
-    "type-check": "tsc"
+    "lint": "expo lint"
   },
   "dependencies": {
-    "@react-native-picker/picker": "~2.11.0",
-    "@react-navigation/native": "~7.1.9",
-    "@react-navigation/native-stack": "~7.3.13",
-    "expo": "^53.0.9",
-    "expo-navigation-bar": "~4.2.4",
+    "@expo/vector-icons": "^14.1.0",
+    "@react-navigation/bottom-tabs": "^7.3.10",
+    "@react-navigation/elements": "^2.3.8",
+    "@react-navigation/native": "^7.1.6",
+    "expo": "~53.0.9",
+    "expo-blur": "~14.1.4",
+    "expo-constants": "~17.1.6",
+    "expo-font": "~13.3.1",
+    "expo-haptics": "~14.1.4",
+    "expo-image": "~2.1.7",
+    "expo-linking": "~7.1.5",
+    "expo-router": "~5.0.6",
+    "expo-splash-screen": "~0.30.8",
     "expo-status-bar": "~2.2.3",
-    "react": "~19.0.0",
-    "react-native": "~0.79.2",
+    "expo-symbols": "~0.4.4",
+    "expo-system-ui": "~5.0.7",
+    "expo-web-browser": "~14.1.6",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
+    "react-native": "0.79.2",
+    "react-native-gesture-handler": "~2.24.0",
+    "react-native-reanimated": "~3.17.4",
     "react-native-safe-area-context": "5.4.0",
-    "react-native-screens": "~4.10.0"
+    "react-native-screens": "~4.10.0",
+    "react-native-web": "~0.20.0",
+    "react-native-webview": "13.13.5",
+    "expo-navigation-bar": "~4.2.4",
+    "@react-native-picker/picker": "2.11.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.27.3",
-    "@eslint/js": "^9.27.0",
+    "@babel/core": "^7.25.2",
     "@types/react": "~19.0.10",
-    "@typescript-eslint/eslint-plugin": "^8.33.0",
-    "@typescript-eslint/parser": "^8.33.0",
-    "eslint": "^9.27.0",
-    "eslint-plugin-react": "^7.37.5",
-    "eslint-plugin-react-native": "^5.0.0",
-    "react-dom": "19.0.0",
-    "react-native-web": "^0.20.0",
     "typescript": "~5.8.3",
-    "typescript-eslint": "^8.33.0"
+    "eslint": "^9.25.0",
+    "eslint-config-expo": "~9.2.0"
   },
   "private": true
 }
