--- src/components/index.ts
@@ -1,4 +1,0 @@
-export { default as HeaderRight } from './HeaderRight';
-export { default as Game } from './Game';
-export { default as Grid } from './Grid';
-export { default as Settings } from './Settings';
