--- src/hooks/index.ts
@@ -1,3 +1,0 @@
-export { default as useGameManager } from './useGameManager';
-export { default as useTimerManager } from './useTimerManager';
-export { GameStatus, TimerStatus } from '../providers';
