--- src/hooks/useGameManager.ts
@@ -1,8 +1,0 @@
-import { useContext } from 'react';
-
-import { GameManagerContext } from '../providers';
-
-
-export default function useGameManager() {
-    return useContext(GameManagerContext);
-}
