--- src/providers/index.ts
@@ -1,2 +1,0 @@
-export { default as GameManagerProvider, GameStatus, GameManagerContext } from './GameManagerProvider';
-export { default as TimerManagerProvider, TimerStatus, TimerManagerContext } from './TimerManagerProvider';
