--- src/styles.ts
@@ -1,15 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-
-// Global stylesheet
-const styles = StyleSheet.create({
-    button: {
-        width: 130,
-        borderRadius: 20,
-        padding: 10,
-        elevation: 2,
-    },
-});
-
-
-export default styles;
