+++ styles/button.ts
@@ -1,0 +1,6 @@
+export default {
+  width: 130,
+  borderRadius: 20,
+  padding: 10,
+  elevation: 2,
+};
