+++ styles/index.ts
@@ -1,0 +1,7 @@
+import { StyleSheet } from 'react-native';
+
+import button from '@/styles/button';
+
+export default StyleSheet.create({
+  button,
+});
